Refactor server API error helper
